requirements.txt
absl-py==0.15.0
accelerate==0.25.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asn1crypto==0.24.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
Automat==0.6.0
backcall==0.2.0
bitsandbytes==0.35.4
cached-property==1.5.2
cachetools==4.2.4
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==3.2.0
chex==0.1.7
click==8.1.7
cloudpickle==1.3.0
colorama==0.4.6
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
configparser==6.0.0
constantly==15.1.0
contourpy==1.1.1
cryptography==2.1.4
cupshelpers==1.0
cycler==0.11.0
decorator==4.4.2
decord==0.6.0
defer==1.0.6
defusedxml==0.7.1
diffusers==0.11.1
dill==0.3.7
distro-info==1.0
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.7.0
etils==1.3.0
executing==2.0.1
filelock==3.13.1
flatbuffers==1.12
flax==0.3.4
fonttools==4.42.1
fsspec==2023.12.0
ftfy==6.1.3
future==0.18.2
gast==0.3.3
gitdb==4.0.10
GitPython==3.1.36
Glances==3.2.6.4
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.9.2
huggingface-hub==0.19.4
hyperlink==17.3.1
idna==2.6
imageio==2.23.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==16.10.1
ipython==8.12.3
jax==0.4.13
jaxlib==0.4.13
jedi==0.19.1
Jinja2==3.1.2
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.4.5
language-selector==0.1
lazy_loader==0.3
libclang==16.0.6
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.3.0
matplotlib-inline==0.1.6
minerl-navigate==1.1.0
ml-dtypes==0.2.0
modelcards==0.1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.6
nest-asyncio==1.5.8
netifaces==0.10.4
networkx==3.1
numpy==1.19.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.4.25
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
optax==0.1.7
orbax==0.1.9
orbax-checkpoint==0.2.3
packaging==21.3
PAM==0.4.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.1
pipdeptree==2.13.0
pkginfo==1.2.1
proglog==0.1.10
promise==2.3
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.21.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.16.1
PyGObject==3.42.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==3.0.9
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.4.1
pyxdg==0.25
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
rich==11.1.0
rsa==4.9
safetensors==0.4.1
scikit-image==0.18.1
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sentry-sdk==1.9.0
service-identity==16.0.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.15.0
smmap==5.0.1
sos==4.3
ssh-import-id==5.7
stack-data==0.6.3
subprocess32==3.5.4
sympy==1.12
system-service==0.3
systemd-python==234
tensorboard==2.4.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-addons==0.13.0
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-gan==2.1.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.12.0
tensorstore==0.1.44
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2023.7.10
tokenizers==0.15.0
toolz==0.12.0
torch==2.1.1+cu118
torchvision==0.13.1
tqdm==4.61.2
traitlets==5.14.0
transformers==4.35.2
triton==2.1.0
twine==1.10.0
Twisted==17.9.0
typeguard==4.1.5
typing_extensions==4.7.1
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
wandb==0.15.11
wcwidth==0.2.12
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
xformers==0.0.23+cu118
xkit==0.0.0
zipp==3.5.0
zope.interface==4.3.2
